Create *****Consecutive Numbers Sum
question tai erokom vaabe approach korbo j ami size check korbo 
that means size 1 er kono answer ache ki nah !
size 2 kono answer ache ki nah!
size 3 er kono answer ache ki nah!

and so on......

evaabe loop ta toto porjonto run jotokhonn nah size er sum amr number taa k cross na kore jaai

LOGIC BEHIND APPROACH:
For a consecutive array of length n, there are several possibilities:
1, 2, 3 ... n
2, 3, 4 ... n, n + 1 → This can also be expressed as 1, 2, 3 ... n, n
3, 4, 5 ... n, n + 1, n + 2 → This can also be expressed as: 1, 2, 3 ... n, n, n
Therefore, we can come into a conclusion that any consecutive array of length n, can be expressed as 1, 2, 3 ... n - 1, n(1 or multiple). If a number S is the sum of a consecutive array of length n, then it must satisfy this property: (S - (1 + n) * n / 2) % n == 0.
We can use the above property to solve this problem. Given a number S, we just need to test if it is the sum of a consecutive array of length 2(we start from 2 because we know that S can be surely made of a consecutive array of length 1, which is S itself), and then we can try if it can be made by a consecutive array of length 3, 4, 5...
One more question is that what will be the max length for the consecutive array that constitutes S? It surely is not of length S. The longest possible length of the consecutive array is of this form: 1 + 2 + 3 + ... + maxLength = S. Therefore, we can stop trying the next length if we find that it will definitely give us a consecutive array whose sum is larger than S.

for better understanding see---> https://youtu.be/oKOCYZd4m7E?t=460